Upgrade Notification service to .NET 6
