+++ src/NotificationService/Events/CustomerRegistered.cs
@@ -1,24 +1,18 @@
-﻿using Pitstop.Infrastructure.Messaging;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace Pitstop.NotificationService.Events;
 
-namespace Pitstop.NotificationService.Events
+public class CustomerRegistered : Event
 {
-    public class CustomerRegistered : Event
-    {
-        public readonly string CustomerId;
-        public readonly string Name;
-        public readonly string TelephoneNumber;
-        public readonly string EmailAddress;
+    public readonly string CustomerId;
+    public readonly string Name;
+    public readonly string TelephoneNumber;
+    public readonly string EmailAddress;
 
-        public CustomerRegistered(Guid messageId, string customerId, string name, string telephoneNumber, string emailAddress) : 
-            base(messageId)
-        {
-            CustomerId = customerId;
-            Name = name;
-            TelephoneNumber = telephoneNumber;
-            EmailAddress = emailAddress;
-        }
+    public CustomerRegistered(Guid messageId, string customerId, string name, string telephoneNumber, string emailAddress) :
+        base(messageId)
+    {
+        CustomerId = customerId;
+        Name = name;
+        TelephoneNumber = telephoneNumber;
+        EmailAddress = emailAddress;
     }
 }
