+++ src/NotificationService/Events/DayHasPassed.cs
@@ -1,14 +1,8 @@
-﻿using Pitstop.Infrastructure.Messaging;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace Pitstop.NotificationService.Events;
 
-namespace Pitstop.NotificationService.Events
+public class DayHasPassed : Event
 {
-    public class DayHasPassed : Event
+    public DayHasPassed(Guid messageId) : base(messageId)
     {
-        public DayHasPassed(Guid messageId) : base(messageId)
-        {
-        }
     }
 }
