+++ src/NotificationService/GlobalUsings.cs
@@ -1,0 +1,18 @@
+global using Pitstop.NotificationService;
+global using Pitstop.Infrastructure.Messaging;
+global using Pitstop.Infrastructure.Messaging.Configuration;
+global using Pitstop.NotificationService.Events;
+global using Pitstop.NotificationService.Model;
+global using Pitstop.NotificationService.NotificationChannels;
+global using Pitstop.NotificationService.Repositories;
+global using Microsoft.Extensions.Hosting;
+global using Newtonsoft.Json.Linq;
+global using Serilog;
+global using System.Text;
+global using Microsoft.Extensions.Configuration;
+global using Microsoft.Extensions.DependencyInjection;
+global using Dapper;
+global using Polly;
+global using System.Data.SqlClient;
+global using System.Net;
+global using System.Net.Mail;
