+++ src/NotificationService/Model/Customer.cs
@@ -1,15 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace Pitstop.NotificationService.Model;
 
-namespace Pitstop.NotificationService.Model
+public class Customer
 {
-    public class Customer
-    {
-        public string CustomerId { get; set; }
-        public string Name { get; set; }
-        public string TelephoneNumber { get; set; }
-        public string EmailAddress { get; set; }
-    }
+    public string CustomerId { get; set; }
+    public string Name { get; set; }
+    public string TelephoneNumber { get; set; }
+    public string EmailAddress { get; set; }
 }
