+++ src/NotificationService/NotificationChannels/IEMailNotifier.cs
@@ -1,12 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Pitstop.NotificationService.NotificationChannels;
 
-namespace Pitstop.NotificationService.NotificationChannels
+public interface IEmailNotifier
 {
-    public interface IEmailNotifier
-    {
-        Task SendEmailAsync(string to, string from, string subject, string body);
-    }
+    Task SendEmailAsync(string to, string from, string subject, string body);
 }
