+++ src/NotificationService/Program.cs
@@ -1,71 +1,32 @@
-﻿using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Pitstop.Infrastructure.Messaging;
-using Pitstop.Infrastructure.Messaging.Configuration;
-using Pitstop.NotificationService.NotificationChannels;
-using Pitstop.NotificationService.Repositories;
-using Serilog;
-using System;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Pitstop.NotificationService
-{
-    class Program
+﻿IHost host = Host
+    .CreateDefaultBuilder(args)
+    .ConfigureServices((hostContext, services) =>
     {
-        public static async Task Main(string[] args)
+        services.UseRabbitMQMessageHandler(hostContext.Configuration);
+
+        services.AddTransient<INotificationRepository>((svc) =>
         {
-            var host = CreateHostBuilder(args).Build();
-            await host.RunAsync();
-        }
+            var sqlConnectionString = hostContext.Configuration.GetConnectionString("NotificationServiceCN");
+            return new SqlServerNotificationRepository(sqlConnectionString);
+        });
 
-        private static IHostBuilder CreateHostBuilder(string[] args)
+        services.AddTransient<IEmailNotifier>((svc) =>
         {
-            var hostBuilder = Host.CreateDefaultBuilder(args)
-                .ConfigureHostConfiguration(configHost =>
-                {
-                    configHost.SetBasePath(Directory.GetCurrentDirectory());
-                    configHost.AddJsonFile("hostsettings.json", optional: true);
-                    configHost.AddJsonFile($"appsettings.json", optional: false);
-                    configHost.AddEnvironmentVariables();
-                    configHost.AddEnvironmentVariables("DOTNET_");
-                    configHost.AddCommandLine(args);
-                })
-                .ConfigureAppConfiguration((hostContext, config) =>
-                {
-                    config.AddJsonFile($"appsettings.{hostContext.HostingEnvironment.EnvironmentName}.json", optional: false);
-                })
-                .ConfigureServices((hostContext, services) =>
-                {
-                    services.UseRabbitMQMessageHandler(hostContext.Configuration);
+            var mailConfigSection = hostContext.Configuration.GetSection("Email");
+            string mailHost = mailConfigSection["Host"];
+            int mailPort = Convert.ToInt32(mailConfigSection["Port"]);
+            string mailUserName = mailConfigSection["User"];
+            string mailPassword = mailConfigSection["Pwd"];
+            return new SMTPEmailNotifier(mailHost, mailPort, mailUserName, mailPassword);
+        });
 
-                    services.AddTransient<INotificationRepository>((svc) =>
-                    {
-                        var sqlConnectionString = hostContext.Configuration.GetConnectionString("NotificationServiceCN");
-                        return new SqlServerNotificationRepository(sqlConnectionString);
-                    });
+        services.AddHostedService<NotificationWorker>();
+    })
+    .UseSerilog((hostContext, loggerConfiguration) =>
+    {
+        loggerConfiguration.ReadFrom.Configuration(hostContext.Configuration);
+    })
+    .UseConsoleLifetime()
+    .Build();
 
-                    services.AddTransient<IEmailNotifier>((svc) =>
-                    {
-                        var mailConfigSection = hostContext.Configuration.GetSection("Email");
-                        string mailHost = mailConfigSection["Host"];
-                        int mailPort = Convert.ToInt32(mailConfigSection["Port"]);
-                        string mailUserName = mailConfigSection["User"];
-                        string mailPassword = mailConfigSection["Pwd"];
-                        return new SMTPEmailNotifier(mailHost, mailPort, mailUserName, mailPassword);
-                    });
-
-                    services.AddHostedService<NotificationManager>();
-                })
-                .UseSerilog((hostContext, loggerConfiguration) =>
-                {
-                    loggerConfiguration.ReadFrom.Configuration(hostContext.Configuration);
-                })
-                .UseConsoleLifetime();
-
-            return hostBuilder;
-        }
-    }
-}
+await host.RunAsync();
